Fix HTML escaping bug in the compiler
This was a weird one. When a template started with a reference, e.g.
"{{foo}} <div>", the reference would be compiled into a SafeBuffer and
the end result would be something like

    ERB::Util.html_escape(presenter.foo) << "<div>"

Since the latter string literal was not marked as HTML safe, it would be
escaped.

This commit changes Compiler such that it has a single SafeBuffer to
which it concatenates all the parts.