Add .exposes_helper to Presenter
This allows delegation of some methods, i.e. root_path, to the
context.  It essentially defines a method on the presenter to
allow passing the method directly to the context.  This not only
allows the method call to bypass the #method_missing method, it
also removes boilerplate code for some methods.