Add extra check in the presenter class resolution
There are a few subclasses of NameError in which "rescue NameError"
also qualifies; namely, NoMethodError also qualifies as a NameError,
and as such, would trigger the rescue block incorrectly.  Since any
class can subclass NameError, the only real option is to raise if
the error isn't an instance of NameError.